#fixes: (refactor) some changes
