+++ packages/mixvisit-lite/src/client-parameters/client-rects/utils.ts
@@ -1,89 +1,9 @@
-/* eslint-disable no-param-reassign */
-
-export type ClientRect = {
-  bottom: number;
-  height: number;
-  left: number;
-  right: number;
-  top: number;
-  width: number;
-  x: number;
-  y: number;
-};
-
 type ElementData = {
   style: string;
   block: string;
 };
 
-export function createClientRectElement(element: ElementData) {
-  if (!element) {
-    throw new Error('No element');
-  }
-
-  const template = document.createElement('template');
-  template.innerHTML = element.block.trim();
-
-  const el = template.content.firstChild as HTMLElement;
-  el.style.cssText = element.style;
-
-  return el.outerHTML;
-}
-
-export async function createElementsAndGetClientRects(
-  iframe: HTMLIFrameElement,
-  iframeWindow: typeof window,
-): Promise<ClientRect[]> {
-  const res: ClientRect[] = [];
-
-  iframe.style.width = '700px';
-  iframe.style.height = '600px';
-
-  const iframeDocument = iframeWindow.document;
-  const { head: iframeHead, body: iframeBody } = iframeDocument;
-
-  // created elements
-  const elementsResult = Object.values(elements)
-    .map(createClientRectElement)
-    .join('');
-
-  const style = iframeDocument.createElement('style');
-  style.textContent = `
-    caption {
-      border: solid 2px darkred;
-      font-size: 20.99px;
-      margin-left: 20.8px;
-    }
-  `;
-  iframeHead.appendChild(style);
-  iframeBody.innerHTML = elementsResult;
-
-  const elementsArr = Array.from(iframeBody.children);
-
-  for (const el of elementsArr) {
-    const {
-      bottom, top,
-      left, right,
-      width, height,
-      x, y,
-    } = el.getBoundingClientRect();
-
-    res.push({
-      bottom,
-      top,
-      left,
-      right,
-      width,
-      height,
-      x,
-      y,
-    });
-  }
-
-  return res;
-}
-
-const elements: Record<string, ElementData> = {
+export const ELEMENTS: Record<string, ElementData> = {
   simpleTest: {
     style: `
       border: solid 2.715px green;
@@ -212,3 +132,17 @@ const elements: Record<string, ElementData> = {
     block: '<div><button type="button"></button></div>',
   },
 };
+
+export function createClientRectElement(element: ElementData) {
+  if (!element) {
+    throw new Error('No element');
+  }
+
+  const template = document.createElement('template');
+  template.innerHTML = element.block.trim();
+
+  const el = template.content.firstChild as HTMLElement;
+  el.style.cssText = element.style;
+
+  return el.outerHTML;
+}
